+++ src/test/java/org/codedefenders/rules/DatabaseRule.java
@@ -1,6 +1,5 @@
-package org.codedefenders;
+package org.codedefenders.rules;
 
-import java.awt.dnd.DnDConstants;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
